refactor: create seperate api handlers per domain. remove async graphql. convert webhooks to openapi.
