+++ api/Cargo.toml
@@ -18,18 +18,15 @@ sea-orm = { version = "^0.10.0", features = [
   "runtime-tokio-rustls",
   "sqlx-postgres",
 ] }
-async-graphql = { version = "5.0.4", features = [
-  "chrono",
-  "uuid",
-  "log",
-  "dataloader",
-  "apollo_tracing",
-] }
 poem = { version = "1.3.50", features = ["anyhow", "test"] }
-async-graphql-poem = "5.0.3"
-poem-openapi = { version = "2.0.23", features = ["swagger-ui", "uuid", "chrono"] }
+poem-openapi = { version = "2.0.23", features = [
+  "swagger-ui",
+  "uuid",
+  "chrono",
+] }
 async-std = { version = "^1", features = ["attributes", "tokio1"] }
 serde = { version = "1.0.152", features = ["derive"] }
+serde_json = { version = "1.0.91" }
 ory-openapi-generated-client = { package = "ory-client", version = "1.1.5" }
 
 [dependencies.hub-core]
