--- api/src/dataloaders/mod.rs
@@ -1,13 +1,0 @@
-pub mod credential;
-pub mod organization;
-pub mod project;
-pub mod project_credential;
-pub mod user;
-pub mod webhook_projects;
-
-pub use credential::Loader as CredentialLoader;
-pub use organization::Loader as OrganizationLoader;
-pub use project::Loader as ProjectLoader;
-pub use project_credential::Loader as ProjectCredentialsLoader;
-pub use user::{MembersLoader, OwnerLoader};
-pub use webhook_projects::Loader as WebhookProjectsLoader;
