--- api/src/dataloaders/organization.rs
@@ -1,37 +1,0 @@
-use std::collections::HashMap;
-
-use async_graphql::{dataloader::Loader as DataLoader, FieldError, Result};
-use poem::async_trait;
-use sea_orm::prelude::*;
-
-use crate::{
-    db::Connection,
-    entities::organizations::{Column, Entity, Model},
-};
-
-#[derive(Debug, Clone)]
-pub struct Loader {
-    pub db: Connection,
-}
-
-impl Loader {
-    #[must_use]
-    pub fn new(db: Connection) -> Self {
-        Self { db }
-    }
-}
-
-#[async_trait]
-impl DataLoader<Uuid> for Loader {
-    type Error = FieldError;
-    type Value = Model;
-
-    async fn load(&self, keys: &[Uuid]) -> Result<HashMap<Uuid, Self::Value>, Self::Error> {
-        let organizations = Entity::find()
-            .filter(Column::Id.is_in(keys.iter().map(ToOwned::to_owned)))
-            .all(self.db.get())
-            .await?;
-
-        Ok(organizations.iter().map(|o| (o.id, o.clone())).collect())
-    }
-}
