+++ api/src/entities/credentials.rs
@@ -1,14 +1,10 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.10.5
 
-use async_graphql::*;
+use poem_openapi::Object;
 use sea_orm::entity::prelude::*;
 
-use super::{organizations::Model as Organization, projects::Model as Project};
-use crate::{dataloaders::project::CredentialId, AppContext};
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, SimpleObject, poem_openapi::Object)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Object)]
 #[sea_orm(table_name = "credentials")]
-#[graphql(concrete(name = "Credential", params()))]
 #[oai(rename = "Credential")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
@@ -23,22 +19,6 @@ pub struct Model {
     pub created_at: DateTime,
     #[oai(read_only)]
     pub created_by: Uuid,
-}
-
-#[ComplexObject]
-impl Model {
-    async fn organization(&self, ctx: &Context<'_>) -> Result<Option<Organization>> {
-        let AppContext {
-            organization_loader,
-            ..
-        } = ctx.data::<AppContext>()?;
-        organization_loader.load_one(self.organization_id).await
-    }
-
-    async fn projects(&self, ctx: &Context<'_>) -> Result<Option<Vec<Project>>> {
-        let AppContext { project_loader, .. } = ctx.data::<AppContext>()?;
-        project_loader.load_one(CredentialId(self.id)).await
-    }
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
