+++ api/src/entities/invites.rs
@@ -1,24 +1,13 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.10.5
 
-use async_graphql::*;
+use poem_openapi::Object;
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
 use super::sea_orm_active_enums::InviteStatus;
 
-#[derive(
-    Clone,
-    Debug,
-    PartialEq,
-    DeriveEntityModel,
-    Eq,
-    SimpleObject,
-    Serialize,
-    Deserialize,
-    poem_openapi::Object,
-)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize, Object)]
 #[sea_orm(table_name = "invites")]
-#[graphql(concrete(name = "Invite", params()))]
 #[oai(rename = "Invite")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
