+++ api/src/entities/members.rs
@@ -1,12 +1,9 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.10.5
 
-use async_graphql::*;
+use poem_openapi::Object;
 use sea_orm::entity::prelude::*;
 
-use super::organizations::Model as Organization;
-use crate::AppContext;
-
-#[derive(Clone, Copy, Debug, PartialEq, DeriveEntityModel, Eq, poem_openapi::Object)]
+#[derive(Clone, Copy, Debug, PartialEq, DeriveEntityModel, Eq, Object)]
 #[sea_orm(table_name = "members")]
 #[oai(rename = "Member", read_only_all)]
 pub struct Model {
@@ -17,48 +14,6 @@ pub struct Model {
     pub created_at: DateTime,
     #[sea_orm(nullable)]
     pub revoked_at: Option<DateTime>,
-}
-
-#[derive(Clone, Copy, SimpleObject, Debug)]
-#[graphql(complex)]
-pub struct Member {
-    pub id: Uuid,
-    pub user_id: Uuid,
-    pub organization_id: Uuid,
-    pub created_at: DateTime,
-    pub revoked_at: Option<DateTime>,
-}
-
-#[ComplexObject]
-impl Member {
-    async fn organization(&self, ctx: &Context<'_>) -> Result<Option<Organization>> {
-        let AppContext {
-            organization_loader,
-            ..
-        } = ctx.data::<AppContext>()?;
-
-        organization_loader.load_one(self.organization_id).await
-    }
-}
-
-impl From<Model> for Member {
-    fn from(
-        Model {
-            id,
-            user_id,
-            organization_id,
-            created_at,
-            revoked_at,
-        }: Model,
-    ) -> Self {
-        Self {
-            id,
-            user_id,
-            organization_id,
-            created_at,
-            revoked_at,
-        }
-    }
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
