+++ api/src/entities/organizations.rs
@@ -1,8 +1,8 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.10.5
-#![allow(clippy::all)]
+
+use poem_openapi::Object;
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
-use poem_openapi::Object;
 
 #[derive(Clone, Debug, PartialEq, Eq, DeriveEntityModel, Serialize, Deserialize, Object)]
 #[sea_orm(table_name = "organizations")]
@@ -17,6 +17,7 @@ pub struct Model {
     pub created_at: DateTime,
     #[oai(read_only)]
     pub deactivated_at: Option<DateTime>,
+    #[oai(read_only)]
     pub svix_app_id: String,
 }
