+++ api/src/entities/owners.rs
@@ -1,12 +1,9 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.10.5
 
-use async_graphql::*;
+use poem_openapi::Object;
 use sea_orm::entity::prelude::*;
 
-use super::organizations::Model as Organization;
-use crate::AppContext;
-
-#[derive(Clone, Debug, Copy, PartialEq, DeriveEntityModel, Eq, poem_openapi::Object)]
+#[derive(Clone, Debug, Copy, PartialEq, DeriveEntityModel, Eq, Object)]
 #[sea_orm(table_name = "owners")]
 #[oai(rename = "Owner", read_only_all)]
 pub struct Model {
@@ -15,45 +12,6 @@ pub struct Model {
     pub user_id: Uuid,
     pub organization_id: Uuid,
     pub created_at: DateTime,
-}
-
-#[derive(SimpleObject, Clone, Debug, Copy)]
-#[graphql(complex)]
-pub struct Owner {
-    pub id: Uuid,
-    pub user_id: Uuid,
-    pub organization_id: Uuid,
-    pub created_at: DateTime,
-}
-
-#[ComplexObject]
-impl Owner {
-    async fn organization(&self, ctx: &Context<'_>) -> Result<Option<Organization>> {
-        let AppContext {
-            organization_loader,
-            ..
-        } = ctx.data::<AppContext>()?;
-
-        organization_loader.load_one(self.organization_id).await
-    }
-}
-
-impl From<Model> for Owner {
-    fn from(
-        Model {
-            id,
-            user_id,
-            organization_id,
-            created_at,
-        }: Model,
-    ) -> Self {
-        Self {
-            id,
-            user_id,
-            organization_id,
-            created_at,
-        }
-    }
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
