+++ api/src/entities/project_credentials.rs
@@ -1,14 +1,11 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.10.5
 
-use async_graphql::*;
+use poem_openapi::Object;
 use sea_orm::entity::prelude::*;
 
-use super::projects::Model as Project;
-use crate::AppContext;
-
-#[derive(Clone, Copy, Debug, PartialEq, DeriveEntityModel, Eq, SimpleObject)]
+#[derive(Clone, Copy, Debug, PartialEq, DeriveEntityModel, Eq, Object)]
 #[sea_orm(table_name = "project_credentials")]
-#[graphql(complex, concrete(name = "ProjectCredential", params()))]
+#[oai(rename = "ProjectCredentials")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub id: Uuid,
@@ -16,14 +13,6 @@ pub struct Model {
     pub project_id: Uuid,
     pub created_at: DateTime,
     pub created_by: Uuid,
-}
-
-#[ComplexObject]
-impl Model {
-    async fn project(&self, ctx: &Context<'_>) -> Result<Option<Project>> {
-        let AppContext { project_loader, .. } = ctx.data::<AppContext>()?;
-        project_loader.load_one(self.project_id).await
-    }
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
