+++ api/src/entities/projects.rs
@@ -1,14 +1,10 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.10.5
 
-use async_graphql::*;
+use poem_openapi::Object;
 use sea_orm::entity::prelude::*;
 
-use super::organizations::Model as Organization;
-use crate::AppContext;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, SimpleObject, poem_openapi::Object)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Object)]
 #[sea_orm(table_name = "projects")]
-#[graphql(complex, concrete(name = "Project", params()))]
 #[oai(rename = "Project")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
@@ -22,17 +18,6 @@ pub struct Model {
     #[sea_orm(nullable)]
     #[oai(read_only)]
     pub deactivated_at: Option<DateTime>,
-}
-
-#[ComplexObject]
-impl Model {
-    async fn organization(&self, ctx: &Context<'_>) -> Result<Option<Organization>> {
-        let AppContext {
-            organization_loader,
-            ..
-        } = ctx.data::<AppContext>()?;
-        organization_loader.load_one(self.organization_id).await
-    }
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
