+++ api/src/entities/sea_orm_active_enums.rs
@@ -1,5 +1,6 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.10.5
 
+use poem_openapi::Enum;
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
@@ -14,7 +15,7 @@ use serde::{Deserialize, Serialize};
     Serialize,
     Deserialize,
     Default,
-    poem_openapi::Enum,
+    Enum,
 )]
 #[sea_orm(rs_type = "String", db_type = "Enum", enum_name = "invitie_status")]
 pub enum InviteStatus {
